fleshed out backend
Added thumbnail to book object and changed everything to JpaRepository

+++ library-app/backend/src/main/java/com/example/library/controller/BookController.java
@@ -23,7 +23,7 @@ public class BookController {
     }
 
     @GetMapping("/{id}")
-    public Book getBookById(@PathVariable int id) {
+    public Book getBookById(@PathVariable Long id) {
         return bookService.getBookById(id);
     }
